% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmat.R
\docType{methods}
\name{+,ggplot,ggplot-method}
\alias{+,ggplot,ggplot-method}
\alias{+,gtable,ggplot-method}
\alias{+,ggplot,gtable-method}
\alias{+,gtable,gtable-method}
\title{Add methods}
\usage{
\S4method{+}{ggplot,ggplot}(e1, e2)

\S4method{+}{gtable,ggplot}(e1, e2)

\S4method{+}{ggplot,gtable}(e1, e2)

\S4method{+}{gtable,gtable}(e1, e2)
}
\arguments{
\item{e1}{A \code{ggplot} or code{gtable} object.}

\item{e2}{A \code{ggplot} or code{gtable} object.}
}
\value{
A \code{gtable} object
}
\description{
Add \code{ggplot} and \code{gtable} objects
}
\examples{
require('ggplot2')

p <- ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = factor(cyl)))
p + p
p + (p + scale_colour_manual(values = c("red","blue", "green")))
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
