% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmat.R
\name{AddEmpty}
\alias{AddEmpty}
\title{Combine empty blocks}
\usage{
AddEmpty(x, leftN, rightN, coreN, reverse = FALSE)
}
\arguments{
\item{x}{A list.}

\item{leftN}{left number.}

\item{rightN}{right number.}

\item{coreN}{center number.}

\item{reverse}{Reverse \code{x} or not.}
}
\value{
A list.
}
\description{
Combine plot elements with empty blocks
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
