% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmat.R
\name{atadd}
\alias{atadd}
\alias{\%@+\%}
\title{Add methods}
\usage{
atadd(e1, e2)

e1 \%@+\% e2
}
\arguments{
\item{e1}{A \code{ggplot} or code{gtable} object.}

\item{e2}{A \code{ggplot} or code{gtable} object.}
}
\value{
A \code{gtable} object
}
\description{
Add \code{ggplot} and \code{gtable} objects
}
\examples{
require('ggplot2')
require('gridExtra')

p <- ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = factor(cyl)))
p \%@+\% p
p \%@+\% grid.arrange(qplot(1,1)) \%@+\% p
p \%@+\% (p + scale_colour_manual(values = c('red', 'blue', 'green')))
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
