% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{ascore}
\alias{ascore}
\title{Convert to the core}
\usage{
ascore(x)
}
\arguments{
\item{x}{A \code{ggplot} or \code{gtable} object.}
}
\value{
A \code{pmat} object.
}
\description{
Convert to the core of a \code{gmat} class object.
}
\examples{
require('ggplot2')
require('magrittr')

qplot(1,1) \%>\% ascore
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
