% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/paraFrame.R
\docType{methods}
\name{Batch}
\alias{Batch}
\alias{Batch,PPIdx-method}
\title{Parallel framework}
\usage{
Batch(x, n, FUN, ...)

\S4method{Batch}{PPIdx}(x, n = 1, FUN, ...)
}
\arguments{
\item{x}{Object with at least profile and linkage indices.}

\item{n}{The number of CPUs or processors.}

\item{FUN}{Functions to process paired linkages.}

\item{...}{Additional parameters passed to \code{FUN}.}
}
\value{
A numeric vector.
}
\description{
Parallel framework for \code{PPIdx} and \code{PPTreeIdx}.
}
\details{
\code{Batch()}: Parallel framework to analysis paired profiles in batch mode.
}
\examples{
require('magrittr')

## Person correlation coefficient
ppBinIdx <- sample(0:1, 10 * 20, replace = TRUE) \%>\% matrix(ncol = 20) \%>\% PP \%>\% PPIdx(1:3, 1:3)
Batch(ppBinIdx, n = 2, cor)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

