% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Batch.R
\docType{methods}
\name{BatchCore}
\alias{BatchCore}
\alias{BatchCore,ANY,matrix-method}
\alias{BatchCore,ANY,big.matrix-method}
\title{Parallel core}
\usage{
BatchCore(p, idx, FUN, ..., n)

\S4method{BatchCore}{ANY,matrix}(p, idx, FUN, ..., n = 1)

\S4method{BatchCore}{ANY,big.matrix}(p, idx, FUN, ..., n = 1)
}
\arguments{
\item{p}{A \code{numeric matrix} or a \code{big.matrix} object.}

\item{idx}{A \code{numeric matrix} object.}

\item{FUN}{Functions to process paired linkages. The \code{FUN} has two parameters, the first parameter is a list \code{list(f = f, t = t, uniID = i)} in which \code{f} and \code{t} is two numeric vector indicating two single profile and \code{uniID} is the current index for parallel programming, and the second parameter is \code{...} used for additional arguments.}

\item{...}{Additional parameters passed to \code{FUN}.}

\item{n}{The number of CPUs or processors.}
}
\value{
A \code{numeric matrix} object.
}
\description{
Parallel core for \code{numeric matrix} and \code{big.matrix}.
}
\details{
Parallel core for the index data.
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
