% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTraits.R
\name{BayesTraitsMode}
\alias{BayesTraitsMode}
\alias{BayesTraitsTest}
\title{Use BayesTraits to test the functional linkages.}
\usage{
BayesTraitsMode(BayesTraitsPath, treeFilePath, binFilePath, method = "ML",
  priorAll = "exp 10", iterNum = 1010000, mode = "Indep", uniID = "1")

BayesTraitsTest(...)
}
\arguments{
\item{BayesTraitsPath}{Full path of the program BayesTraits.}

\item{treeFilePath}{Full path of tree file, and the the tree should be rooted and in nexus format.}

\item{binFilePath}{Full path of phylogenetic profiles, and only contains two genes. The speperate character by tab. If the paired genes are all "1" or all "0", then 0 will be returned.}

\item{method}{Either 'MCMC' or 'ML' (Maximum Likelihood) method.}

\item{priorAll}{Prior distribution.}

\item{iterNum}{Iteration number only for 'MCMC' method, the default value is 1010000.}

\item{mode}{Either 'Indep' for independent and 'Dep' for dependent.}

\item{uniID}{Unique ID for output file, and it is designed for parallel programming.}

\item{...}{Inherit parameters from the "BayesTraitsMode()" in this package.}
}
\value{
\code{BayesTraitsMode()}: Harmonic mean.

\code{BayesTraitsTest()}: A list.
}
\description{
BayesTraits Mode
}
\details{
\code{BayesTraitsMode()}: Test the functinal linkages by BayesTraits using independent or dependent mode.

\code{BayesTraitsTest()}: Test the a pair of functional linkage.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
