% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BayesTraitsBatch.R
\docType{methods}
\name{BayesTraits}
\alias{BayesTraits}
\alias{BayesTraits,PPTreeIdx-method}
\title{Batch process of BayesTraits}
\usage{
BayesTraits(x, ..., n = 1)

\S4method{BayesTraits}{PPTreeIdx}(x, ..., n = 1)
}
\arguments{
\item{x}{\code{PPTreeIdx} object.}

\item{...}{Additional parameters
\itemize{
  \item \code{BayesTraitsPath}: Full path of the program \href{http://www.evolution.reading.ac.uk/BayesTraits.html}{BayesTraits}.
  \item \code{method}: Either "MCMC" or "ML" (Maximum Likelihood) method. "ML" is set as default.
}
Parameters only used in \code{MCMC}:
\itemize{
  \item \code{priorAll}: Prior distribution.
  \item \code{iterNum}:  Iteration number with default value 1010000.
}}

\item{n}{The number of CPUs or processors.}
}
\value{
A \code{PPResult} object.
}
\description{
BayesTraits
}
\details{
A wrapper of the BayesTraits test of paired profiles. Please download and install \href{http://www.evolution.reading.ac.uk/BayesTraits.html}{BayesTraits (version 3.0.1)} before.
}
\examples{
require('magrittr')
require('ape')

tree <- system.file('extdata', 'bioinfoTree.nex', package = "PhyloProfileKit") \%>\% read.nexus
ppPath <- system.file('extdata', 'bioinfoProfile.csv', package = "PhyloProfileKit")

sceP <- ppPath \%>\% read.csv(row.names = 1) \%>\% as.matrix \%>\% PP
scePTI <- sceP \%>\% PPTree(tree) \%>\% PPTreeIdx(1:6, 1:6)

\dontrun{
## replace "BTPath" with the full path of BayesTraits in your system, for example in Linux/OS
## BTPath <- '/path/to/BayesTraitsV2/BayesTraitsV2'
## or in Windows
## BTPath <- '/path/to/BayesTraitsV2/BayesTraitsV2.exe'
BayesTraits(scePTI, BayesTraitsPath = BTPath, n = 2)
}

}
\references{
\href{http://www.evolution.reading.ac.uk/BayesTraits.html}{BayesTraits program}

\url{https://www.ncbi.nlm.nih.gov/pubmed/?term=17090580}

\url{https://www.ncbi.nlm.nih.gov/pubmed/?term=16103904}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
