% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PP.R
\docType{methods}
\name{rbind2,PP,ANY-method}
\alias{rbind2,PP,ANY-method}
\alias{cbind2,PP,ANY-method}
\title{rbind or cbind PP objects}
\usage{
\S4method{rbind2}{PP,ANY}(x, y, ...)

\S4method{cbind2}{PP,ANY}(x, y, ...)
}
\arguments{
\item{x}{A \code{PP} object .}

\item{y}{A \code{PP} object or a named numeric matrix.}

\item{...}{Additional parameters.}
}
\value{
A \code{PP} object.
}
\description{
Row bind or column bind of a \code{PP} object
}
\details{
\code{rbind(..., deparse.level = 1)}: Row bind a \code{PP} object with another \code{PP} object or a named numeric matrix.

\code{cbind(..., deparse.level = 1)}: Column bind a \code{PP} object with another \code{PP} object or a named numeric matrix.
}
\examples{
require('magrittr')
ppC <- matrix(rnorm(10 * 20),
              ncol = 20,
              dimnames = list(paste0('protein', 1:10),
                              paste0('spe', 1:20))) \%>\% PP

ppM <- matrix(rnorm(10 * 20),
              ncol = 20,
              dimnames = list(paste0('protein', 11:20),
                              paste0('spe', 21:40)))

cbind(ppC, ppM)
cbind(ppC, PP(ppM))
rbind(ppC, ppM)
rbind(ppC, PP(ppM))
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
