% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SimDistBatch.R
\docType{methods}
\name{ChooseSimDistFun}
\alias{ChooseSimDistFun}
\alias{ChooseSimDistFun,PP-method}
\title{Choose SimDist function}
\usage{
ChooseSimDistFun(x, method, ...)

\S4method{ChooseSimDistFun}{PP}(x, method, ...)
}
\arguments{
\item{x}{A \code{PP} object.}

\item{method}{A character string.
\itemize{
  \item \code{"SimCor"}: Person's correlation coefficient.
  \item \code{"SimJaccard"}: Jaccard similarity.
  \item \code{"SimMI"}: Mutual information.
  \item \code{"DistHamming"}: Hamming distance.
  \item \code{"DistEuclidean"}: Euclidean distance.
}

\code{"SimCor"} and \code{"DistHamming"} are not recommended for binning profiles.}

\item{...}{Additional parameters passed to \code{method}.
\itemize{
  \item \code{bin}: Integer. The number of breaks in the \code{"MI"} method for continuous profiles.
}}
}
\value{
A function.
}
\description{
Similarity or distance function.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
