% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CollapseTree}
\alias{CollapseTree}
\title{Collapse tree}
\usage{
CollapseTree(edgeMat, tipNum, f, t)
}
\arguments{
\item{edgeMat}{A edge mat could be generated from the "ape" package.The first row should be (root --> nodes).}

\item{tipNum}{A int value. Tip (species) number.}
}
\value{
A numeric matrix with four columns: first two are collapsed edges and last two represent collapsed two profiles.
}
\description{
Collapse phylogenetic profiles according to the phylogenetic tree
}
\details{
The branches with same profile pattern are merged.
}
\references{
\href{http://rsif.royalsocietypublishing.org/content/5/19/151}{collapse tree description}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
