% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{comb2_internal}
\alias{comb2_internal}
\alias{combWhole_internal}
\alias{combSelf_internal}
\title{Combination}
\usage{
comb2_internal(x)

combWhole_internal(x, y, self = FALSE, bidirect = FALSE)

combSelf_internal(x)
}
\arguments{
\item{x}{Vector.}

\item{y}{Another vector, and every element of \code{x} should be in \code{y}.}

\item{self}{Whether include self pairs.}

\item{bidirect}{Whether to include two directions.}
}
\value{
\code{comb2_internal()}/\code{combWhole_internal()}/\code{combSelf_internal()}: Matrix with 2 columns
}
\description{
Combination functions
}
\details{
\code{comb2_interal()}: Pair combination of the input vector \code{x}.

\code{combWhole_internal()}: Pair combination of the input vector \code{x} and \code{y}.

\code{combSelf_internal()}: Self pairs.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
