% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPResult.R
\docType{methods}
\name{Compare,PPResult,numeric-method}
\alias{Compare,PPResult,numeric-method}
\title{Compare methods}
\usage{
\S4method{Compare}{PPResult,numeric}(e1, e2)
}
\arguments{
\item{e1}{A \code{PPResult} object.}

\item{e2}{A numeric value.}
}
\value{
A \code{logic vector}
}
\description{
Compare \code{PPResult} objects with a numeric value.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
