% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InferGainNodes}
\alias{CountRepeatIdx}
\alias{DolloDist}
\alias{InferEdge}
\alias{InferGainNodes}
\alias{MergeList}
\title{Dollo's parsimony distance}
\usage{
InferGainNodes(gainList)

InferEdge(edgeMat, tipPath, pr)

DolloDist(edgeMat, tipPath, pr1, pr2)

MergeList(x)

CountRepeatIdx(idx, y)
}
\arguments{
\item{gainList}{A list of ancestors of each gain tips. In each elements, the first one is the root.}

\item{edgeMat}{A edge mat could be generated from the "ape" package.The first row should be (root --> nodes).}

\item{tipPath}{A list of ancestors of each gain tips. In each elements, the first one is the root.}

\item{pr}{A numeric vector indicates the "presence-absence" pattern."pr" should be in the same order with the tips of tree}

\item{pr1}{1st "presence-absence" pattern.}

\item{pr2}{2nd "presence-absence" pattern.}

\item{x}{A list only contains unsigned interger vectors.}

\item{idx}{Index}

\item{y}{A unsigned integer vector}
}
\value{
InferGainNodes(): A vector (with tips) of 1 and 0.

DolloDist(): An integer.

A numeric edge present and absent matrix.

A number indicating the Dollo's parsimony distance.

A merged numeric vector.

Repeat number of y[idx] in y.
}
\description{
Use Dollo's parsimony distance to evaluate similarity patterns.
}
\details{
Algorithm:
1. Find the latest common ancestors (LCA) for all gain tips.
2. Select the nodes from each tip to the LCA.

InferGainNodes(): Infer the present nodes and tips.

DolloDist(): Infer the Dollo's parsimony distance.
}
\examples{
## example tree
library('ape')
testTreeText <- '((((t1, t2),(t3, t4)),t5), (t6, (t7, t8)));'
testTree <- read.tree(text = testTreeText)
plot(testTree)

## node path
pathList <- nodepath(testTree)
InferGainNodes(pathList[1])
InferGainNodes(pathList[1:3])

InferEdge(testTree$edge, pathList, c(0, 0, 0, 0, 1, 1, 1, 1))

DolloDist(testTree$edge, pathList, c(0, 0, 0, 0, 1, 1, 1, 1), c(0, 0, 0, 0, 0, 1, 1, 1))
DolloDist(testTree$edge, pathList, c(1, 1, 1, 1, 1, 0, 0, 1), c(0, 0, 0, 0, 1, 0, 0, 1))
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pubmed/?term=17535793}
}
\keyword{internal}

