% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DolloBatch.R
\name{DolloDistBatch}
\alias{DolloDistBatch}
\title{Batch process of Dollo's parsimony distance.}
\usage{
DolloDistBatch(ftMat, profileMat, edgeMat, tipPath, n = 1)
}
\arguments{
\item{ftMat}{A two column matrix which should at least have rownames.}

\item{profileMat}{The phylogenetic profile data with 1 and 0 denoting the presence and absence of orthologous, respectively. It is a named numeric matrix, columns are genes and rows are species.}

\item{edgeMat}{A edge mat could be generated from the "ape" package.The first row should be (root --> nodes).}

\item{tipPath}{A list of ancestors of each gain tips. In each elements, the first one is the root.}

\item{n}{Number of CPUs.}
}
\value{
A numeric vector.
}
\description{
Dollo's parsimony distance of input phylogenetic profiles
}
\details{
DolloDistBatch(): Dollo's parsimony distance in batch mode.
}
\examples{
require('ape')

descPath <- system.file('extdata', 'bioinfoTree.nex', package = "PhyloProfile")
tree <- read.nexus(descPath)
pathList <- nodepath(tree)

pair1 <- system.file('extdata', 'bioinfoGenepair1.txt', package = "PhyloProfile")
oneP <- read.table(pair1, sep = '\\t', row.names = 1)
sampleP <- matrix(sample(0:1, nrow(oneP) * 1000, replace = TRUE), ncol = 1000)
colnames(sampleP) <- paste0('gene', 1:1000)
rownames(sampleP) <- rownames(oneP)
ft <- matrix(paste0('gene', sample(1:1000, 10)), ncol = 2)
rownames(ft) <- paste0('p_', 1:nrow(ft))

DolloDistBatch(ftMat = ft, profileMat = sampleP, edgeMat = tree$edge, tipPath = pathList, n = 2)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\seealso{
dollo
}

