% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotElements.R
\name{EmptyEle}
\alias{EmptyEle}
\title{Empty block}
\usage{
EmptyEle(...)
}
\arguments{
\item{...}{ggplot2 geom_point() additional parameters}
}
\value{
ggplot2 object
}
\description{
Empty white block
}
\details{
A empty white block used as interval to organize plot
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}

