% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_tree.R
\name{ExtractSeg}
\alias{ExtractSeg}
\title{Extract segments}
\usage{
ExtractSeg(d)
}
\arguments{
\item{d}{Data structure from tree from the ggtree package.}
}
\value{
A data frame.
}
\description{
Extract segment data
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
