% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDef.R
\docType{data}
\name{fatp}
\alias{fatp}
\title{Phylogenetic profiles of human F1Fo ATP synthase subunits.}
\format{A list}
\description{
A list contains:
\itemize{
  \item \code{atpPhylo}: A numeric matrix representing the phylogenetic profiles of F1Fo ATP synthase subuints.
  \item \code{atpPhyloBit}: A numeric matrix representing the BLASTP bit scores of F1Fo ATP synthase subuints.
  \item \code{complex}: Fo and F1 complex subunits.
  \item \code{complexCol}: Colours for complex subunits.
  \item \code{doamin}: Domain of species.
  \item \code{domainCol}: Colours of domains.
}
}
\references{
Unpublished data from Yulong Niu
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
