% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDef.R
\docType{data}
\name{fatp1}
\alias{fatp1}
\title{Phylogenetic profiles of human F1Fo ATP synthase subunits.}
\format{A list}
\description{
A list containing three objects:
"atpPhylo" is a numeric matrix representing the phylogenetic profiles of F1Fo ATP synthase subuints.
"atpPhyloBit" is a numeric matrix representing the BLASTP bit scores of F1Fo ATP synthase subuints.
"specol" is the vector of colors corresponding to the species in "atpPhylo".
"genecol" is the vector of colors corresponding to the genes in "atpPhylo".
"domain" is the domain names corresponding to the "speCol"
}
\references{
Unpublished data from Yulong Niu
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
