% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotElements.R
\name{geom_emptyblock}
\alias{geom_emptyblock}
\title{Empty block}
\usage{
geom_emptyblock(...)
}
\arguments{
\item{...}{ggplot2 geom_point() additional parameters}
}
\value{
ggplot2 object
}
\description{
Empty white block
}
\details{
A empty white block used as interval to organize plot
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
