% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotElements.R
\name{geom_legend}
\alias{geom_legend}
\title{Retrieve ggplot2 legend}
\usage{
geom_legend(g, ...)
}
\arguments{
\item{g}{ggplot2 object}

\item{...}{ggplot2 theme() additional parameters mainly about control the legend position}
}
\value{
Legend object
}
\description{
ggplot legend
}
\details{
Retrieve the ggplot legend from a given ggplot2 object. This function may cause a blank display device, a dirty method is used to walkaround.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
\url{https://github.com/hadley/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs}

\url{https://stackoverflow.com/questions/17012518/why-does-this-r-ggplot2-code-bring-up-a-blank-display-device}

\url{https://github.com/hadley/ggplot2/issues/809}
}
\keyword{internal}

