% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{gg2list}
\alias{gg2list}
\title{Transfer to list.}
\usage{
gg2list(x)
}
\arguments{
\item{x}{A \code{ggplot}/\code{gtable} object.}
}
\value{
A list.
}
\description{
Transfer to a list
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
