% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmat.R
\name{gg2t}
\alias{gg2t}
\title{Transfer to \code{gtable} objects.}
\usage{
gg2t(x)
}
\arguments{
\item{x}{A \code{ggplot} or \code{gtable} object.}
}
\value{
A \code{gtable} object.
}
\description{
Transfer to \code{gtable}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
