% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{gmat-class}
\alias{gmat-class}
\title{This class represents the data structure of matrix-like plot.}
\description{
This class represents the data structure of matrix-like plot.
}
\section{Slots}{

\describe{
\item{\code{core}}{core plot.}

\item{\code{left}}{left plot.}

\item{\code{right}}{right plot.}

\item{\code{top}}{top plot.}

\item{\code{bottom}}{bottom plot.}
}}

\author{
Yulong Niu \email{niuylscu@gmail.com}
}
