% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/PPIdx.R
\docType{methods}
\name{IdxData}
\alias{IdxData}
\alias{IdxData<-}
\alias{IdxData,PPIdx-method}
\alias{IdxData<-,PPIdx,matrix-method}
\alias{IdxData<-,PPIdx,big.matrix-method}
\title{Select and replace \code{PPIdx}/\code{PPTreeIdx} data}
\usage{
IdxData(x, ...)

IdxData(x, ...) <- value

\S4method{IdxData}{PPIdx}(x, ...)

\S4method{IdxData}{PPIdx,matrix}(x, ...) <- value

\S4method{IdxData}{PPIdx,big.matrix}(x, ...) <- value
}
\arguments{
\item{x}{A \code{PPIdx}/\code{PPTreeIdx} object.}

\item{...}{Additional parameters.}

\item{value}{A numeric matrix.}
}
\value{
\code{PPData(x)}: A numeric matrix.

\code{PPData(x) <- value}: An update \code{PPIdx}/\code{PPTreeIdx} object.
}
\description{
Select and replace the indices of a \code{PPIdx}/\code{PPTreeIdx} object
}
\details{
\code{IdxData(x)}: Extract the indices from a \code{PPIdx}/\code{PPTreeIdx} object.

\code{IdxData(x) <- value}: Replace the indices of a \code{PPIdx}/\code{PPTreeIdx} object.
}
\examples{
require('magrittr')
ppContinuous <- matrix(rnorm(10 * 20),
                       ncol = 20,
                       dimnames = list(paste0('protein', 1:10),
                                       paste0('spe', 1:20))) \%>\% PP
ppi <- PPIdx(ppContinuous, 1:3, 1:3)

## extract indices
IdxData(ppi)

## replace indices
imat <- matrix(sample(1:10, 2 * 3, replace = FALSE),
               ncol = 2,
               dimnames = list(paste0('link', 1:3),
                               c('from', 'to')))
IdxData(ppi)  <- imat
## type is changed to "binning"
ppContinuous

}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
