% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isTwoRowsEqual}
\alias{isTwoRowsEqual}
\title{Compare two rows Test}
\usage{
isTwoRowsEqual(m)
}
\arguments{
\item{m}{A numeric matrix with two rows.}
}
\value{
logic value.
}
\description{
Test two rows are equal
}
\details{
Pairwise comparison tow rows
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
