% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmat.R
\name{Lay}
\alias{Lay}
\alias{LaySide}
\title{Construct the layout matrix}
\usage{
Lay(x)

LaySide(x, side = "top")
}
\arguments{
\item{x}{A \code{gmat} object.}
}
\value{
A numeric matrix
}
\description{
Construct the layout matrix of a \code{gmat} object.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
