% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotProfile.R
\name{legend_spe}
\alias{legend_spe}
\title{Phylo legend}
\usage{
legend_spe(speCol, classCol, ...)
}
\arguments{
\item{speCol}{A vector of colors with names of species, which are the same as colnames of "phyloData" (may not in the same order).}

\item{classCol}{A vector of colors with the names defining the species categories.}

\item{...}{Parameters from geom_legend}
}
\value{
ggplot2 object
}
\description{
Legend of species in phylogenetic profilings
}
\details{
Legend annotating the species categories, like the class or phylum.
}
\examples{
data(fatp)
speLeg <- legend_spe(fatp$specol, fatp$domain, legend.position = 'left')
\dontrun{
# plot legend
require(gridExtra)
grid.arrange(speLeg, ncol = 1)
}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

