% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmat.R
\name{\%@<\%}
\alias{\%@<\%}
\alias{\%@>\%}
\alias{\%@^\%}
\alias{\%@v\%}
\title{Add locations}
\usage{
x \%@<\% y

x \%@>\% y

x \%@^\% y

x \%@v\% y
}
\arguments{
\item{x}{A \code{pmat} object.}

\item{y}{A \code{ptable} object.}
}
\value{
A \code{pmat} object.
}
\description{
Add plot object to \code{pmat}.
}
\details{
%@<%: at left.

%@>%: at right.

%@^%: at top.

%@v%: at bottom.
}
\examples{
require('gridExtra')
require('ggplot2')

p <- qplot(1,1)
new('gmat') \%@<\% grid.arrange(p)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
