% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Norm.R
\docType{methods}
\name{Norm}
\alias{Norm}
\alias{Norm,PP,character-method}
\title{Normalization phylogenetic profile}
\usage{
Norm(x, method, ...)

\S4method{Norm}{PP,character}(x, method, ...)
}
\arguments{
\item{x}{A \code{PP} object.}

\item{method}{A character string, "NPP" or "SVD". Method used to normalize raw bit score phylogenetic profile.}

\item{...}{Additional parameters passed to normalization methods. Parameters shared in \code{NPP} and \code{SVD}:
\itemize{
  \item \code{bitCutoff}: Cutoff of minimum bit score.
  \item \code{bitReset}: Reset bit scores lower than the \code{bitCutoff}.
  \item \code{minConserve}: Minimum number of homologous. The proteins with homologous less than this value are discarded.
}
Parameters only used in \code{SVD}:
\itemize{
  \item \code{trimming}: Top percentages of the unitary matrix, for example 0.3 means top 30\% columns are kept.
}}
}
\value{
A \code{PP} object.
}
\description{
Normalization methods
}
\details{
Normalization of raw bit score profile by Singular Value Decomposition (SVD) or Normalized Phylogenetic Profile (NPP).

The SVD is retrieved from the \href{https://bitbucket.org/andrea/svd-phy}{SVD-Phy package} with enhanced performance. The NPP is firstly described in the \href{http://www.nature.com/nature/journal/v493/n7434/full/nature11779.html}{paper}.
}
\examples{
require("magrittr")
ppRawBit <- matrix(sample(0:200, 2000 * 1000, replace = TRUE),
                   ncol = 1000,
                   dimnames = list(paste0('protein', 1:2000),
                                   paste0('spe', 1:1000))) \%>\% PP

## SVD
Norm(ppRawBit, method = 'SVD', bitCutoff = 100, bitReset = 0, minConserve = 12, trimming = 0.8)

## NPP
Norm(ppRawBit, method = 'NPP', bitCutoff = 60, bitReset = 1, minConserve = 12)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
\href{https://bitbucket.org/andrea/svd-phy}{SVD-Phy package}

\href{http://bioinformatics.oxfordjournals.org/content/suppl/2015/11/25/btv696.DC1/SVD-Phy-supplementary-material.docx}{SVD description}

\href{http://www.nature.com/nature/journal/v493/n7434/extref/nature11779-s1.pdf}{NPP description}
}

