% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SVDNorm}
\alias{SVDNorm}
\alias{SVDPhy}
\title{Singular value decomposition normalization of bit score matrix}
\usage{
SVDNorm(rawBitM, bitCutoff, bitReset, minConserve, trimming)

SVDPhy(bitM, bitReset, minConserve, trimming)
}
\arguments{
\item{rawBitM}{Raw bit score matrix.}

\item{bitCutoff}{Minimum value of the bit score.}

\item{bitReset}{Reset the bit score for ones lower than the `bitCutoff`.}

\item{minConserve}{Minimum number of homologous. The proteins with homologous less than this value are discarded.}

\item{trimming}{A percentages top unitary matrix.}

\item{bitM}{Bit score matrix, for example the BLASTP or STRING bit scores. It is a named numeric matrix, columns are species and rows are genes.}
}
\value{
SVDNorm(): SVD normalized bit score matrix.

SVDPhy(): A L^2 normalized unitary matrix.
}
\description{
SVD normalization
}
\details{
Algorithm:

Step1: rawBitM < bitCutoff to bitReset;

Step2: In each row (species), x/max(x);

Step3: L^2 SVD normalization.

The core SVD normalization is retrieved from the SVD-Phy package with performance modification.
}
\references{
\href{https://bitbucket.org/andrea/svd-phy}{SVD-Phy package}

\href{http://bioinformatics.oxfordjournals.org/content/suppl/2015/11/25/btv696.DC1/SVD-Phy-supplementary-material.docx}{SVD description}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
