% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotElements.R
\name{phylo_legend}
\alias{phylo_legend}
\title{Phylo legend}
\usage{
phylo_legend(classCol, ...)
}
\arguments{
\item{classCol}{A vector of colors with the names defining the species categories.}

\item{...}{Parameters from geom_legend}
}
\value{
ggplot2 object
}
\description{
Legend of phylogenetic profilings
}
\details{
Legend annotating the species categories, like the class or phylum.
}
\examples{
data(fatp)

phylo_legend(fatp$domain)
phyloVec <- c('#00C19F', '#93AA00', '#D39200')
names(phyloVec) <- c('Eukaryota', 'Archaea', 'Bacteria')
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

