% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmat.R
\docType{methods}
\name{plot,gmat,missing-method}
\alias{plot,gmat,missing-method}
\title{Plot methods}
\usage{
\S4method{plot}{gmat,missing}(x, y, ...)
}
\arguments{
\item{x}{A \code{gmat}}

\item{y}{Not set ("missing").}

\item{...}{Parameters passed to \code{grid.arrange()} in the gridExtra package.}
}
\value{
A \code{gtable} object.
}
\description{
Plot method for \code{gmat} objects
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
