% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfile.R
\name{PlotPhyloProfile}
\alias{PlotPhyloProfile}
\title{Plot phylogenetic profiles}
\usage{
PlotPhyloProfile(phyloData, geneNameSize = 3, geneNameCol = "grey55",
  geneBetweenBlockCol = NA, presentCol = "steelblue",
  absentCol = "grey91", speCol, geneCol, classCol, widthsShinkage = c(0.9,
  0.7, 0.3, 7, 2.2), heightsShinkage = c(0.3, 7), ...)
}
\arguments{
\item{phyloData}{The phylogenetic profile data with 1 and 0 denoting the presence and absence of orthologous, respectively. The "phyloData" should be a numeric matrix, of which the row is gene and column is species. The "phyloData"has row names and column names which will be used for the dendrogram of row and column.}

\item{geneNameSize}{The size of gene names label, and the default value is 3.}

\item{geneNameCol}{The colour of gene names, and the default value is "grey55".}

\item{geneBetweenBlockCol}{The space color between gene blocks, and the default value is "NA" meaning no space color. If the number of genes is samll, for example less than 20, setting it as 'white' is fine.}

\item{presentCol}{The color of present 1, the default value is "steelblue".}

\item{absentCol}{The color of present 0, the default value is "grey91".}

\item{speCol}{A vector of colors with names of species, which are the same as colnames of "phyloData" (may not in the same order).}

\item{geneCol}{A vector of colors with names of genes, which are the same as rownames of "phyloData" (may not in the same order).}

\item{classCol}{A vector of colors with the names defining the species categories.}

\item{widthsShinkage}{The shinkage width vector.}

\item{heightsShinkage}{The shinkage width vector.}

\item{...}{Parameters from pp_legend}
}
\value{
A plot object.
}
\description{
Plot phylogenetic profiles with gene and species clusters.
}
\details{
A combination plot of phylogenetic profiles with genes and species clusters.
}
\examples{
data(fatp)
ATPphyloPlot <- PlotPhyloProfile(fatp$atpPhylo, speCol = fatp$specol, geneCol = fatp$genecol,
classCol = fatp$domain, legend.position = 'left')
\dontrun{
# an example of saving output figures
cairo_pdf('FATPprofilePlot.pdf')
ATPphyloPlot <- PlotPhyloProfile(fatp$atpPhylo, speCol = fatp$specol, geneCol = fatp$genecol,
classCol = fatp$domain, legend.position = 'left')
dev.off()
}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
