% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotprofile.R
\docType{methods}
\name{plotprofile}
\alias{plotprofile}
\alias{plotprofile,PP-method}
\title{Wrapped profiles plot function.}
\usage{
plotprofile(x, method, ...)

\S4method{plotprofile}{PP}(x, method, ...)
}
\arguments{
\item{x}{\code{PP}/\code{PPTreeIdx} object.}

\item{method}{Distance method used in the \code{dist()}.}

\item{...}{Additional parameters
\itemize{
  \item \code{proSize}: A numeric value. Size of protein names.
  \item \code{proGroup}: A factor indicating protein groups.
  \item \code{proGroupCol}: A named character vector indicating protein group colour.
  \item \code{speGroup}: A factor indicating species groups.
  \item \code{speGroup}: A named character vector indicating species group colour.
}}
}
\value{
A \code{gtable} object.
}
\description{
Plot phylogenetic profile with additional data
}
\details{
\itemize{
  \item x is a \code{PP} or \code{PPIdx} object: \code{method} works for both dimensions (proteins and species).
  \item x is a \code{PPTreeIdx} object: \code{method} works only for rows (proteins).
}
}
\examples{
data(fatp)

plotprofile(PP(fatp$atpPhylo), method = NA)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
