% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_text.R
\name{pp_htext}
\alias{pp_htext}
\alias{pp_vtext}
\title{pp plot text}
\usage{
pp_htext(x, shift = 0.5, ...)

pp_vtext(x, shift = 0.5, ...)
}
\arguments{
\item{x}{A character vector}

\item{shift}{A numeric value indicating shit scale.}

\item{...}{Parameters passed to \code{geom_text()} in the ggplot2 package.}
}
\value{
A \code{gg} class object
}
\description{
Plot text
}
\details{
Plot text, for example column names and row names.

\code{pp_htext()}: Plot text in the horizontal direction.

\code{pp_vtext()}: Plot text in the vertical direction.
}
\examples{
pp_htext(letters[1:10], colour = 'grey55', size = 3)
pp_vtext(rep(LETTERS[1:2], 3), colour = factor(rep(c('blue', 'red'), 3)))
}
\seealso{
\code{\link[ggplot2]{geom_text}}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
