% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_empty.R
\name{pp_empty}
\alias{pp_empty}
\title{pp plot empty}
\usage{
pp_empty(...)
}
\arguments{
\item{...}{Parameters passed to \code{geom_point()} in the ggplot2 package.}
}
\value{
A \code{gg} class object
}
\description{
Plot empty block
}
\details{
Plot empty blocks, for example an empty white block.
}
\examples{
require(ggplot2)

pp_empty(colour = 'white')
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
