% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_legend.R
\name{pp_legend}
\alias{pp_legend}
\title{Retrieve ggplot2 legend}
\usage{
pp_legend(g, ...)
}
\arguments{
\item{g}{\code{ggplot2} object.}

\item{...}{Parameters passed to \code{theme()} in the ggplot2 package.}
}
\value{
A \code{gtable} object.
}
\description{
ggplot legend
}
\details{
Retrieve the ggplot legend from a given ggplot2.
}
\examples{
require('ggplot')

p <- ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = factor(cyl)))
plot(pp_legend(p, legend.position = 'left'))
}
\references{
\url{https://github.com/hadley/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs}

\url{https://stackoverflow.com/questions/17012518/why-does-this-r-ggplot2-code-bring-up-a-blank-display-device}

\url{https://github.com/hadley/ggplot2/issues/809}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
