% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_profile.R
\name{pp_profile}
\alias{pp_profile}
\title{pp plot profiles}
\usage{
pp_profile(x, legend.position = "none", ...)
}
\arguments{
\item{x}{A numeric matrix.}

\item{legend.position}{Position of the legend and the default is "none" (no legend). See the \code{theme()} in the ggplot2 package.}

\item{...}{Parameters passed to \code{geom_tile()} in the ggplot2 package.}
}
\value{
A \code{gg} class object
}
\description{
Plot profiles
}
\details{
Plot a phylogenetic profile, which is a matrix.
}
\examples{
require('ggplot2')
require('magrittr')

## a binning profile
ppB <- sample(0:1, 5 * 20, replace = TRUE) \%>\% matrix(nrow = 5)
pp_profile(ppB) + scale_fill_manual(values = c('grey91', 'steelblue'))

## a continuous profile
ppC <- rnorm(5 * 20) \%>\% matrix(nrow = 5)
pp_profile(ppC) + scale_fill_gradient2()
}
\seealso{
\code{\link[ggplot2]{geom_tile}}

\code{\link[ggplot2]{theme}}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
