% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_tile.R
\name{pp_htile}
\alias{pp_htile}
\alias{pp_vtile}
\title{pp plot tiles}
\usage{
pp_htile(x, ...)

pp_vtile(x, ...)
}
\arguments{
\item{x}{A character vector}

\item{...}{Parameters passed to \code{geom_tile()} in the ggplot2 package.}
}
\value{
A \code{gg} class object
}
\description{
Plot tiles
}
\details{
Plot tiles to discrete groups

\code{pp_htext()}: Plot tiles in the horizontal direction.

\code{pp_vtext()}: Plot tiles in the vertical direction.
}
\examples{
pp_htile(rep(letters[1:3], 4))
pp_vtile(rep(LETTERS[1:2], 3)) + scale_fill_manual(values = c('red','blue', 'green'))
}
\seealso{
\code{\link[ggplot2]{geom_tile}}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
