% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_tree.R
\name{pp_tree}
\alias{pp_tree}
\title{pp plot tree}
\usage{
pp_tree(x, ...)
}
\arguments{
\item{x}{A tree object}

\item{...}{Parameters passed to \code{geom_tile()} in the ggplot2 package.}
}
\value{
A \code{gg} class object
}
\description{
Plot trees
}
\details{
Plot phylogenetic trees and dendrograms
}
\examples{
require('ggplot2')

pp_tile(rep(letters[1:3], 4))
pp_tile(rep(LETTERS[1:2], 3)) + scale_fill_manual(values = c('red','blue', 'green'))
pp_tile(rep(letters[1:3], 4)) + coord_flip()
}
\seealso{
\code{\link[ggplot2]{geom_tile}}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
