% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{PPData}
\alias{PPData}
\alias{PPData<-}
\title{Select and replace PP data}
\usage{
PPData(x, ...)

PPData(x, ...) <- value
}
\arguments{
\item{x}{A PP object.}

\item{...}{Additional parameters}

\item{value}{A numeric matrix.}
}
\value{
\code{PPData(x)}: A numeric matrix.

\code{PPData(x) <- value}: An update PP object.
}
\description{
Select and replace generic methods
}
\details{
\code{PPData(x)}: Extract the data matrix from a PP object.

\code{PPData(x) <- value}: Replace the data matrix of a PP object.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

