% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PP.R
\docType{methods}
\name{PPData,PP-method}
\alias{PPData,PP-method}
\alias{PPData<-,PP,matrix-method}
\title{Select and replace PP data}
\usage{
\S4method{PPData}{PP}(x, ...)

\S4method{PPData}{PP,matrix}(x, ...) <- value
}
\arguments{
\item{x}{A PP object.}

\item{...}{Additional parameters}

\item{value}{A numeric matrix.}
}
\value{
A numeric matrix.

\code{PPData(x)}: A numeric matrix.

\code{PPData(x) <- value}: An update PP object.
}
\description{
Select and replace the data matrix of a PP object
}
\details{
\code{PPData(x)}: Extract the data matrix from a PP object.

\code{PPData(x) <- value}: Replace the data matrix of a PP object.
}
\examples{
ppContinuous <- PP(matrix(rnorm(10 * 20),
                          ncol = 20,
                          dimnames = list(paste0('protein', 1:10),
                                          paste0('spe', 1:20))))
## extract data matrix
PPData(ppContinuous)

## replace whole data matrix
rmat <- matrix(sample(0:1, 20 * 30, replace = TRUE),
               ncol = 30,
               dimnames = list(paste0('pro', 1:20),
                               paste0('spe', 1:30)))
PPData(ppContinuous)  <- rmat
## type is changed to "binning"
ppContinuous

}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

