% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PPIdx-class}
\alias{PPIdx-class}
\title{This class represents the data structure of phylogenetic profile with linkage indices.}
\description{
This class represents the data structure of phylogenetic profile with linkage indices.
}
\section{Slots}{

\describe{
\item{\code{idx}}{An integer matrix with two columns. It validates the rownames and colnames of the profile matrix.}
}}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

