% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPIdx.R
\name{PPIdx}
\alias{PPIdx}
\title{Constructor of \code{PPIdx}}
\usage{
PPIdx(value, p)
}
\arguments{
\item{value}{A character matrix with two columns. The linkages containing proteins not in \code{p} (profile) are removed.}

\item{p}{A \code{PP} object}
}
\value{
A \code{PPIdx} object
}
\description{
The constructor the \code{PPIdx} class
}
\details{
Construct a \code{PP} object from a character matrix.
}
\examples{
require('magrittr')

## construct a PP object without dimnames
ppBinning <- sample(0:1, 10 * 20, replace = TRUE) \%>\% matrix(ncol = 20) \%>\% PP
linkM <- sample(1:30, 20 * 2, replace = TRUE) \%>\% paste0('protein', .) \%>\% matrix(ncol = 2)
ppBinIdx <- PPIdx(linkM, ppBinning)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

