% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PPResult-class}
\alias{PPResult-class}
\title{This class represents the data structure of phylogenetic profiling results.}
\description{
This class represents the data structure of phylogenetic profiling results.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A numeric vector representing the profiling data.}

\item{\code{idx}}{An integer matrix with two columns. It validates the rownames and colnames of the profile.}

\item{\code{pnames}}{A character vector storing rownames (protein names) of the profile.}

\item{\code{method}}{A character string indicating the profiling method.}
}}

\author{
Yulong Niu \email{niuylscu@gmail.com}
}
