% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRING.R
\name{PPSTRING}
\alias{PPSTRING}
\title{Constructor from the STRING database}
\usage{
PPSTRING(speID, idx = 1:10, symbets = FALSE, STRINGnorm = FALSE, n = 1,
  verbose = FALSE)
}
\arguments{
\item{speID}{Number. NCBI taxonomy identifier must be present in the STRING database. Use \code{get_STRING_species()} from the STRINGdb package to determine supported organisms.}

\item{idx}{A numeric vector or "all".}

\item{symbets}{Whether to keep only symmetrical best hits.}

\item{STRINGnorm}{Whether you use the normalized or raw bit scores.}

\item{n}{The number of CPUs or processors, and the default value is 1.}

\item{verbose}{Whether to print retrieve process.}
}
\value{
A \code{PP} object.
}
\description{
The constructor of the \code{PP} class from the STRING database
}
\details{
Retrieve the bit score matrix from a give organism. The default bit score threshold in STRING is 50, homologous with values smaller than 50 are not shown. For these homologous, the default value is set as 0.
}
\examples{
## human top 5 proteins
PPSTRING(9606, idx = 1:5, n = 2)

\dontrun{
## all human proteins
PPRCSTRING(9606, idx = "all", n = 4)
}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
