% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PPTree-class}
\alias{PPTree-class}
\title{This class represents the data structure of phylogenetic profile with phylogenetic tree.}
\description{
This class represents the data structure of phylogenetic profile with phylogenetic tree.
}
\section{Slots}{

\describe{
\item{\code{tree}}{A \code{phylo} object indicating the phylogenetic tree from the \code{ape} package.}
}}

\seealso{
The raw definition of \code{\link[ape]{phylo}}.
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
