% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPTree.R
\name{PPTree}
\alias{PPTree}
\title{Constructor of \code{PPTree}}
\usage{
PPTree(x, tree)
}
\arguments{
\item{x}{A numeric matrix.}

\item{tree}{A \code{phylo} object.}
}
\value{
A \code{PPTree} object.
}
\description{
The constructor the \code{PPTree} class
}
\details{
Construct a \code{PPTree} object from a numeric matrix.
}
\examples{
require('magrittr')
require('ape')

ppMat <- sample(0:1, 10 * 20, replace = TRUE) \%>\% matrix(ncol = 20)
tree <- rtree(8, tip.label = paste0('spe', 8:1))
PPTree(ppMat, tree)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
