% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PPTreeIdx-class}
\alias{PPTreeIdx-class}
\title{This class represents the data structure of the phylogenetic profile with linkage indices and a phylogenetic tree.}
\description{
This class represents the data structure of the phylogenetic profile with linkage indices and a phylogenetic tree.
}
\section{Slots}{

\describe{
\item{\code{tree}}{A numeric vector representing the profiling data.}
}}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\seealso{
The raw definition of \code{\link[ape]{phylo}}.
}

