% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPTreeIdx.R
\name{PPTreeIdx}
\alias{PPTreeIdx}
\title{Constructor of \code{PPTreeIdx}}
\usage{
PPTreeIdx(pidx, tree)
}
\arguments{
\item{pidx}{A \code{PPIdx} object.}

\item{tree}{A \code{phylo} object.}
}
\value{
A \code{PPTreeIdx} object.
}
\description{
The constructor the \code{PPTreeIdx} class
}
\details{
Construct a \code{PP} object. The species (columns) are in the same order of phylogenetic tree tips, and for the species not in the tree are deleted.
}
\examples{
require('magrittr')
require('ape')

ppBinning <- sample(0:1, 10 * 20, replace = TRUE) \%>\% matrix(ncol = 20) \%>\% PP
ppLink <- sample(1:30, 20 * 2, replace = TRUE) \%>\% paste0('protein', .) \%>\% matrix(ncol = 2) \%>\% PPIdx(ppBinnning, .)

ppTree <- rtree(10, tip.label = paste0('spe', 10:1))
PPTreeIdx(ppLink, ppTree)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
