% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotprofile.R
\name{ProfileCore}
\alias{ProfileCore}
\title{Core plot of phylogenetic profile}
\usage{
ProfileCore(p, proSize = 3, proGroup = NA, proGroupCol = NA,
  speGroup = NA, speGroupCol = NA)
}
\arguments{
\item{p}{A integer matrix (binning profile).}

\item{proSize}{A numeric value. Size of protein names.}

\item{proGroup}{A factor indicating protein groups.}

\item{proGroupCol}{A named character vector indicating protein group colour.}

\item{speGroup}{A factor indicating species groups.}

\item{speGroupCol}{A named character vector indicating species group colour.}
}
\value{
A \code{gmat} object.
}
\description{
The core plot of phylogenetic profile
}
\details{
Plot of phylogenetic profile with protein names, a protein colour bar (optional), and a species colour bar (optional).
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
