% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PP.R
\docType{methods}
\name{[,PP-method}
\alias{[,PP-method}
\alias{[<-,PP-method}
\title{Select or replace PP objects}
\usage{
\S4method{[}{PP}(x, i, j, ..., drop = TRUE)

\S4method{[}{PP}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A \code{PP} object.}

\item{i, j, ...}{Indices.}

\item{drop}{Whether the result is coerced to the lowest possible dimension. Please always set it as \code{FALSE}.}

\item{value}{A number, numeric vector or numeric matrix.}
}
\value{
A \code{PP} object.
}
\description{
Select or replace parts of a \code{PP} object
}
\details{
\code{x[i, j, ..., drop]}: Select parts of a \code{PP} object.

\code{x[i, j, ..., drop] <- value}: Replace parts of a \code{PP} object.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

