% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PP.R, R/PPIdx.R, R/PPResult.R
\docType{methods}
\name{show,PP-method}
\alias{show,PP-method}
\alias{show,PPIdx-method}
\alias{show,PPResult-method}
\title{Show methods}
\usage{
\S4method{show}{PP}(object)

\S4method{show}{PPIdx}(object)

\S4method{show}{PPResult}(object)
}
\arguments{
\item{object}{A \code{PP}/\code{PPIdx} object.}
}
\value{
Show messages.
}
\description{
Show method for \code{PP} and \code{PPIdx} objects
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

