% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/simDist.R
\name{SimJaccard}
\alias{DistHamming}
\alias{SimCor}
\alias{SimJaccard}
\alias{SimMI}
\title{similarity and distance}
\usage{
SimJaccard(pairProfile)

SimCor(pairProfile)

DistHamming(pairProfile)

SimMI(pairProfile)
}
\arguments{
\item{pairProfile}{A paired phylogenetic profile, columns are genes and rows are species.}
}
\value{
A numeric value.
}
\description{
Similarity or distance of paired phylogenetic profile
}
\details{
SimCor(): Person's correlation coefficient.
SimJaccard(): Jaccard similarity.
SimMI(): Mutual information.
DistHamming(): Hamming distance.
}
\examples{
## alpha and beta subunits from the F-type ATP synthase.
data(fatp)
ab <- t(fatp$atpPhylo[c('ATP5A1', 'ATP5B'), ])

## Person's correlation coefficient
corAB <- SimCor(ab)
## Jaccard similarity
jacAB <- SimJaccard(ab)
## Mutual information
MIAB <- SimMI(ab)
## Hamming distance
hamAB <- DistHamming(ab)

}
\author{
Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}
}
\seealso{
SimDistBatch
}

