% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDistFun.R
\name{SimDistBatch}
\alias{SimDistBatch}
\title{Batch process of similarity and distance.}
\usage{
SimDistBatch(ftMat, profileMat, FUN, n = 1)
}
\arguments{
\item{ftMat}{A two column matrix which should at least have rownames.}

\item{profileMat}{The phylogenetic profile data with 1 and 0 denoting the presence and absence of orthologous, respectively. It is a named numeric matrix, columns are species anniuyulongd rows are genes.}

\item{FUN}{Functions to calculate single similarity or distance.}

\item{n}{The number of CPUs or processors, and the default value is 1.}
}
\value{
A numeric vector
}
\description{
Similarity or distance of input phylogenetic profiles
}
\details{
SimDistBatch(): Similarity and distance in batch mode.
}
\examples{
data(fatp)
f1 <- t(combn(rownames(fatp$atpPhylo)[1:6], 2))
SimDistBatch(f1, fatp$atpPhylo, SimCor, 2)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\seealso{
simdist
}

