% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVD.R
\name{SVDNor}
\alias{SVDNor}
\alias{SVDPhy}
\title{Singular value decomposition normalization of bit score matrix}
\usage{
SVDNor(rawBitM, bitCutoff = 60, bitReset = 1, trimming = 0.3,
  minConserve = -0.1)

SVDPhy(bitM, trimming, minConserve)
}
\arguments{
\item{rawBitM}{Raw bit score matrix.}

\item{bitCutoff}{Minimum value of the bit score.}

\item{bitReset}{Reset the bit score for ones lower than the `bitCutoff`.}

\item{trimming}{A percentages top unitary matrix.}

\item{minConserve}{Minimum number of homologous in each species. The species with homologous less than thsi value are discarded.}

\item{bitM}{Bit score matrix, for example the BLASTP or STRING bit scores. It is a named numeric matrix, columns are species and rows are genes.}
}
\value{
SVDNor(): SVD normalized bit score matrix.

SVDPhy(): A L^2 normalized unitary matrix.
}
\description{
SVD normalization
}
\details{
Algorithm:

Step1: rawBitM < bitCutoff to bitReset;

Step2: In each row (species), x/max(x);

Step3: L^2 SVD normalization.

The core SVD normalization is retrieved from the SVD-Phy package with performance modification.
}
\examples{
data(fatp)
svdM <- SVDNor(fatp$atpPhyloBit)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
\url{http://bioinformatics.oxfordjournals.org/content/suppl/2015/11/25/btv696.DC1/SVD-Phy-supplementary-material.docx}

\url{https://bitbucket.org/andrea/svd-phy}
}
\seealso{
\code{\link{NPPNor}}
}

