% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotProfile.R
\name{theme_phylo}
\alias{theme_phylo}
\title{Blank theme for plots}
\usage{
theme_phylo(...)
}
\arguments{
\item{...}{ggplot2 theme() additional parameters}
}
\value{
ggplot2 object
}
\description{
Phylo theme
}
\details{
A totally blank gpplot2 theme
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}

