% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotElements.R
\name{theme_pp}
\alias{theme_pp}
\title{Blank theme}
\usage{
theme_pp(...)
}
\arguments{
\item{...}{ggplot2 theme() additional parameters}
}
\value{
ggplot2 object
}
\description{
PhyloProfile theme
}
\details{
A totally blank gpplot2 theme expect for legend
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}

