% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utilities.R
\name{Phylo2Mat}
\alias{Phylo2Mat}
\alias{Phylo2MatY}
\alias{InitM}
\alias{Phylo2MatX}
\title{tree utilities}
\usage{
Phylo2Mat(x)

Phylo2MatY(x)

InitM(x)

Phylo2MatX(x)
}
\arguments{
\item{x}{A \code{phylo} class}
}
\value{
\code{Phylo2Mat()}: A data.frame indicating x and y locations.

\code{Phylo2MatX()}: A data.frame indicating x locations.

\code{Phylo2MatY()}: A data.frame indicating y locations.

\code{InitM()}: A numeric matrix indicating edge and edge length
}
\description{
Deal with \code{phylo} class objects for plot
}
\details{
\code{Phylo2Mat()}: x and y locations.

\code{Phylo2MatX()}: x locations.

\code{Phylo2MatY()}: y locations.

\code{InitM()}: edge and edge length.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
