% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isBinMat_internal}
\alias{isBinMat_internal}
\alias{validMatNames_}
\alias{validTreeMat_}
\title{utilities}
\usage{
isBinMat_internal(x)

validMatNames_(x, warnName)

validTreeMat_(m, t)
}
\arguments{
\item{x}{A matrix.}

\item{warnName}{A character string indicating the class names.}

\item{m}{A numeric matrix.}

\item{t}{A phylo tree.}

\item{x}{A numeric matrix.}
}
\value{
\code{isBinMat_internal()}: Logic.

\code{valiMatNames_()}: \code{TRUE} or a warning message.

\code{valiTreeMat_()}: \code{TRUE} or a warning message.
}
\description{
Additional utilities
}
\details{
\code{isBinMat_internal()}: Whether a \code{numeric matrix} is binning.

\code{valiMatNames_()}: Validate a \code{numeric matrix} with row names and column names.

\code{valiTreeMat_()}: Validate a \code{numeric matrix} with trees.
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
