% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{atadd}
\alias{atadd}
\alias{\%@+\%}
\title{Add methods}
\usage{
atadd(e1, e2)

e1 \%@+\% e2
}
\arguments{
\item{e1}{A \code{ggplot} or list.}

\item{e2}{A \code{ggplot} or list.}
}
\value{
A list.
}
\description{
Connect \code{ggplot} objects to a list.
}
\examples{
require('ggplot2')

p <- ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = factor(cyl)))
p \%@+\% p
p \%@+\% qplot(1,1) \%@+\% p
p \%@+\% (p + scale_colour_manual(values = c('red', 'blue', 'green')))
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
