% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Batch.R
\docType{methods}
\name{Batch}
\alias{Batch}
\alias{Batch,PPIdx-method}
\title{Parallel framework}
\usage{
Batch(x, FUN, ..., n)

\S4method{Batch}{PPIdx}(x, FUN, ..., n = 1)
}
\arguments{
\item{x}{Object with at least profile and linkage indices.}

\item{FUN}{Functions to process paired linkages. The \code{FUN} has two parameters, the first parameter is a list \code{list(f = f, t = t, uniID = i)} in which \code{f} and \code{t} is two numeric vector indicating two single profile and \code{uniID} is the current index for parallel programming, and the second parameter is \code{...} used for additional arguments.}

\item{...}{Additional parameters passed to \code{FUN}.}

\item{n}{The number of CPUs or processors.}
}
\value{
A numeric vector.
}
\description{
Parallel framework for \code{PPIdx} and \code{PPTreeIdx}.
}
\details{
\code{Batch()}: Parallel framework to analysis paired profiles in batch mode.
}
\examples{
require('magrittr')

## Person correlation coefficient
ppBinIdx <- sample(0:1, 10 * 20, replace = TRUE) \%>\% matrix(ncol = 20) \%>\% PP \%>\% PPIdx(1:3, 1:3)
testfun <- function(eachArg, ...) {sum(eachArg$f * eachArg$t)}
Batch(ppBinIdx, testfun, n = 2)
}
\seealso{
Batch process of similarity/distance \code{\link{SimDist}}.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
