% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTraitsBatch.R
\name{BayesTraitsMode}
\alias{BayesTraitsMode}
\alias{BayesTraitsTest}
\title{Use BayesTraits to test the functional linkages.}
\usage{
BayesTraitsMode(BayesTraitsPath, treeFilePath, binFilePath, method = "MCMC",
  priorAll = "exp 10", Stones = "100 1000", mode = "Indep", uniID = "1")

BayesTraitsTest(...)
}
\arguments{
\item{BayesTraitsPath}{Full path of the program BayesTraits.}

\item{treeFilePath}{Full path of tree file, and the the tree should be rooted and in nexus format.}

\item{binFilePath}{Full path of phylogenetic profiles, and only contains two genes. The speperate character by tab. If the paired genes are all "1" or all "0", then 0 will be returned.}

\item{method}{Either "MCMC" (Markov chain Monte Carlo) or "ML" (maximum likelihood) method.}

\item{priorAll}{Prior distribution.}

\item{Stones}{Stone sampler only for "MCMC" method, the default value is "100 1000" meaning 100 stones and 1000 iterations per stone to estimate the marginal likelihood.}

\item{mode}{Either "Indep" for independent and "Dep" for dependent.}

\item{uniID}{Unique ID for output file, and it is designed for parallel programming.}

\item{...}{Inherit parameters from the "BayesTraitsMode()" in this package.}
}
\value{
\code{BayesTraitsMode()}: Log marginal likelihood for the method \code{MCMC} and log likelihood for the method \code{ML}.

\code{BayesTraitsTest()}: A list.
}
\description{
BayesTraits Mode
}
\details{
\code{BayesTraitsMode()}: Test the functinal linkages by BayesTraits using independent or dependent mode.

\code{BayesTraitsTest()}: Test the a pair of functional linkage.
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
