% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTraits.R
\name{BayesTraitsMode}
\alias{BayesTraits}
\alias{BayesTraitsBatch}
\alias{BayesTraitsMode}
\title{Use BayesTraits to test the functional linkages.}
\usage{
BayesTraitsMode(BayesTraitsPath, treeFilePath, binFilePath, method = "ML",
  priorAll = "exp 10", iterNum = 1010000, mode = "Indep", uniID = "1")

BayesTraits(...)

BayesTraitsBatch(ftMat, profileMat, n = 1, ...)
}
\arguments{
\item{BayesTraitsPath}{Full path of the program BayesTraits.}

\item{treeFilePath}{Full path of tree file, and the the tree should be rooted and in nexus format.}

\item{binFilePath}{Full path of phylogenetic profiles, and only contains two genes. The speperate character by tab. If the paired genes are all "1" or all "0", then 0 will be returned.}

\item{method}{Either 'MCMC' or 'ML' (Maximum Likelihood) method.}

\item{priorAll}{Prior distribution.}

\item{iterNum}{Iteration number only for 'MCMC' method, the default value is 1010000.}

\item{mode}{Either 'Indep' for independent and 'Dep' for dependent.}

\item{uniID}{Unique ID for output file, and it is designed for parallel programming.}

\item{...}{Inherit parameters from the "BayesTraitsMode()" in this package.}

\item{ftMat}{A two column matrix which should at least have rownames.}

\item{profileMat}{The phylogenetic profile data with 1 and 0 denoting the presence and absence of orthologous, respectively. It is a named numeric matrix, columns are genes and rows are species.}

\item{n}{Number of CPUs.}
}
\value{
BayesTraitsMode(): Harmonic mean.

BayesTraits(): A list.

BayesTraitsBatch(): A list.
}
\description{
BayesTraits
}
\details{
BayesTraitsMode(): Test the functinal linkages by BayesTraits using independent or dependent mode.

BayesTraits(): Test the a pair of functional linkage.

BayesTraitsBatch(): Test a batch of functional linkages
}
\examples{

descPath <- system.file('extdata', package = "PhyloProTree")

tree <- file.path(descPath, 'bioinfoTree.nex')
pair00 <- file.path(descPath, 'bioinfoGenepair00.txt')
pair11 <- file.path(descPath, 'bioinfoGenepair11.txt')
pair1 <- file.path(descPath, 'bioinfoGenepair1.txt')

BTtree <- file.path(descPath, 'Primates.trees')
BTpair <- file.path(descPath, 'Primates.txt')

oneP <- read.table(pair1, sep = '\\t', row.names = 1)
sampleP <- matrix(sample(0:1, nrow(oneP) * 1000, replace = TRUE), ncol = 1000)
colnames(sampleP) <- paste0('gene', 1:1000)
rownames(sampleP) <- rownames(oneP)
ft <- matrix(paste0('gene', sample(1:1000, 10)), ncol = 2)
rownames(ft) <- paste0('p_', 1:nrow(ft))

\dontrun{
## replace "BTPath" with the full path of BayesTraits in your computer
## all absence in the profile
BayesTraits(BTPath, tree, pair00)
## all presence in the profile
BayesTraits(BTPath, tree, pair11)
## a nomal profile
BayesTraits(BTPath, tree, pair1)

## examples from BayesTraits
BayesTraits(BTPath, BTtree, BTpair)
BayesTraits(BTPath, BTtree, BTpair, method = 'MCMC')

## batch
BayesTraitsBatch(ftMat = ft, profileMat = sampleP, n = 2,
BayesTraitsPath = BTPath, treeFilePath = BTtree)
}

}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
\url{http://www.evolution.reading.ac.uk/BayesTraits.html}

\url{https://www.ncbi.nlm.nih.gov/pubmed/?term=17090580}

\url{https://www.ncbi.nlm.nih.gov/pubmed/?term=16103904}
}

