% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BayesTraits.R
\docType{methods}
\name{BayesTraits}
\alias{BayesTraits}
\alias{BayesTraits,PPTreeIdx-method}
\title{Batch process of BayesTraits}
\usage{
BayesTraits(x, ..., n)

\S4method{BayesTraits}{PPTreeIdx}(x, ..., n = 1)
}
\arguments{
\item{x}{\code{PPTreeIdx} object.}

\item{...}{Additional parameters
\itemize{
  \item \code{BayesTraitsPath}: Full path of the program \href{BayesTraits}{http://www.evolution.reading.ac.uk/BayesTraits.html}.
  \item \code{method}: Either "MCMC" or "ML" (Maximum Likelihood) method. "ML" is set as default.
}
Parameters if \code{method} is set as \code{"MCMC"}:
\itemize{
  \item \code{priorAll}: Prior distribution.
  \item \code{iterNum}:  Iteration number with default value 1010000.
}}

\item{n}{The number of CPUs or processors.}
}
\value{
A \code{PPResult} object.
}
\description{
BayesTraits
}
\details{
A wrapper of the BayesTraits test of paired profiles. Please download and install \href{BayesTraits}{http://www.evolution.reading.ac.uk/BayesTraits.html}.
}
\examples{
require('magrittr')
require('ape')

tree <- system.file('extdata', 'bioinfoTree.nex', package = "PhyloProfile") \%>\% read.nexus
ppPath <- system.file('extdata', 'bioinfoProfile.csv', package = "PhyloProfile")

sceP <- ppPath \%>\% read.csv(row.names = 1) \%>\% as.matrix \%>\% PP
sceT <- PPIdx(sceP, 1:6, 1:6) \%>\% PPTreeIdx(tree)
\dontrun{
## replace "BTPath" with the full path of BayesTraits in your system, for example in Linux/OS
## \\code{BTPath <- '/program/BayesTraitsV2/BayesTraitsV2'}
## or in Windows
## \\code{BTPath <- '/program/BayesTraitsV2/BayesTraitsV2.exe'}
BayesTraits(sceT, BayesTraitsPath = BTPath, n = 2)
}

}
\references{
\href{http://www.evolution.reading.ac.uk/BayesTraits.html}{BayesTraits program}

\url{https://www.ncbi.nlm.nih.gov/pubmed/?term=17090580}

\url{https://www.ncbi.nlm.nih.gov/pubmed/?term=16103904}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
