% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CollapseTree}
\alias{CollapseTree}
\title{Collapse tree}
\usage{
CollapseTree(edgeMat, tipNum, f, t)
}
\arguments{
\item{edgeMat}{A edge mat could be generated from the "ape" package.The first row should be (root --> nodes).}

\item{tipNum}{A int value. Tip (species) number.}

\item{f}{Numeric vector indicating a gene profile.}

\item{t}{Numeric vector indicating a gene profile.}
}
\value{
A numeric matrix with two columns representing collapsed two profiles.
}
\description{
Collapse phylogenetic profiles by the phylogenetic tree
}
\details{
The branches with same profile pattern are merged.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
\href{http://rsif.royalsocietypublishing.org/content/5/19/151}{collapse tree description}
}
\keyword{internal}

