% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InferGainNodes}
\alias{InferGainNodes}
\alias{InferEdge}
\alias{DolloDist}
\alias{MergeList}
\title{Dollo's parsimony distance}
\usage{
InferGainNodes(gainList)

InferEdge(edgeMat, tipPath, pr)

DolloDist(edgeMat, tipPath, f, t)

MergeList(x)
}
\arguments{
\item{gainList}{A list of ancestors of each gain tips. In each elements, the first one is the root.}

\item{edgeMat}{A edge mat could be generated from the "ape" package.The first row should be (root --> nodes).}

\item{tipPath}{A list of ancestors of each gain tips. In each elements, the first one is the root.}

\item{pr}{A numeric vector indicates the "presence-absence" pattern."pr" should be in the same order with the tips of tree}

\item{x}{A list only contains unsigned interger vectors.}
}
\value{
\code{InferGainNodes()}: A vector (with tips) of 1 and 0.

\code{DolloDist()}: An integer.

\code{InferEdge()}: A numeric edge present and absent matrix.

A number indicating the Dollo's parsimony distance.

A merged numeric vector.
}
\description{
Use Dollo's parsimony distance to evaluate similarity patterns.
}
\details{
Algorithm:
1. Find the latest common ancestors (LCA) for all gain tips.
2. Select the nodes from each tip to the LCA.

\code{InferGainNodes()}: Infer the present nodes and tips.

\code{DolloDist()}: Infer the Dollo's parsimony distance.

\code{InferEdge()}: Infer the present and absent of each edges.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pubmed/?term=17535793}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
