% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotProfile.R
\docType{data}
\name{fatp}
\alias{fatp}
\title{Phylogenetic profiles of human F1Fo ATP synthase subunits.}
\format{A list}
\description{
A list containing three objects:
"atpPhylo" is a numeric matrix representing the phylogenetic profiles of F1Fo ATP synthase subuints.
"specol" is the vector of colors corresponding to the species in "atpPhylo".
"genecol" is the vector of colors corresponding to the genes in "atpPhylo".
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
Unpublished data from Yulong Niu
}

