% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/PPIdx.R
\docType{methods}
\name{Idx}
\alias{Idx}
\alias{Idx,PP,matrix-method}
\alias{Idx,PP,big.matrix-method}
\alias{Idx,PP,numeric-method}
\title{Index linkages}
\usage{
Idx(p, x, ...)

\S4method{Idx}{PP,matrix}(p, x, ...)

\S4method{Idx}{PP,big.matrix}(p, x, ...)

\S4method{Idx}{PP,numeric}(p, x, ...)
}
\arguments{
\item{p}{A \code{PP} object.}

\item{x}{A matrix with two columns or a numeric vector. Proteins not in \code{p} (profile) are removed. The numeric vector indicates the indices of proteins.
\itemize{
  \item a character matrix: Each row containing the protein names is a linkage.
  \item a numeric matrix or big.matrix: Each row containing the protein indices is a linkage.
  \item a numeric vector: Each element is the index of a interested protein.
}}

\item{...}{Additional parameters if \code{x} is a numeric vector.
\itemize{
  \item \code{y}: Another numeric vector used to generate paired linkages with \code{x}. Each element of \code{x} should be in \code{y}.
  \item \code{self}: Whether include self pairs, and default set is \code{FALSE}.
  \item \code{bidirect}: Whether to include two directions, and default set is \code{FALSE}.
}}
}
\value{
A \code{numeric matrix} object.
}
\description{
Index linkages methods
}
\details{
Index linkages from protein IDs or indices in batch.
}
\examples{
require('magrittr')

ppBinning <- sample(0:1, 10 * 20, replace = TRUE) \%>\% matrix(ncol = 20) \%>\% PP

## pre-built linkages
linkM <- sample(1:30, 20 * 2, replace = TRUE) \%>\% paste0('protein', .) \%>\% matrix(ncol = 2)
Idx(ppBinning, linkM)

## top 3 proteins with whole profiles
Idx(ppBinning, 1:3, 1:nrow(ppBinning))
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
