% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Profile.R
\name{legend_spe}
\alias{legend_spe}
\title{Phylo legend}
\usage{
legend_spe(classCol, ...)
}
\arguments{
\item{classCol}{A vector of colors with the names defining the species categories.}

\item{...}{Parameters from pp_legend}
}
\value{
ggplot2 object
}
\description{
Legend of species in phylogenetic profiling
}
\details{
Legend annotating the species categories, like the class or phylum.
}
\examples{
data(fatp)
speLeg <- legend_spe(fatp$domain, legend.position = 'left')
\dontrun{
# plot legend
require(gridExtra)
grid.arrange(speLeg, ncol = 1)
}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
