% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{atleft}
\alias{atleft}
\alias{atright}
\alias{attop}
\alias{atbottom}
\alias{\%@<\%}
\alias{\%@>\%}
\alias{\%@^\%}
\alias{\%@v\%}
\title{Add locations}
\usage{
atleft(x, y)

atright(x, y)

attop(x, y)

atbottom(x, y)

x \%@<\% y

x \%@>\% y

x \%@^\% y

x \%@v\% y
}
\arguments{
\item{x}{A \code{pmat} object.}

\item{y}{A \code{ptable} or \code{ggplot} object.}
}
\value{
A \code{pmat} object.
}
\description{
Add plot object to \code{pmat}.
}
\details{
\code{atleft()}: at left.

\code{atright()}: at right.

\code{attop()}: at top.

\code{atbottom()}: at bottom.
}
\examples{
require('ggplot2')

p <- qplot(1,1)
new('gmat') \%@<\% p
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
