% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCorMat.R
\name{PlotPhyloCor}
\alias{PlotPhyloCor}
\title{Plot correlation matrix}
\usage{
PlotPhyloCor(phyloData, gradientCol = colorRampPalette(rev(brewer.pal(n = 7,
  name = "RdYlBu")))(100), geneNameSize = 3, geneNameCol = "grey55",
  geneBetweenBlockCol = NA, geneCol, showCorVal = TRUE,
  widthsShinkage = c(0.9, 0.7, 0.3, 7), heightsShinkage = c(7, 0.7))
}
\arguments{
\item{phyloData}{The phylogenetic profile data with 1 and 0 denoting the presence and absence of orthologous, respectively. The "phyloData" should be a numeric matrix, of which the row is gene and column is species.The "phyloData"has row names and column names which will be used for the dendrogram of row and column.}

\item{gradientCol}{The gradien colours for correlation matrix.}

\item{geneNameSize}{The size of gene names label, and the default value is 3.}

\item{geneNameCol}{The colour of gene names, and the default value is "grey55".}

\item{geneBetweenBlockCol}{The space color between gene blocks, and the default value is "NA" meaning no space color. If the number of genes is samll, for example less than 20, setting it as 'white' is fine.}

\item{geneCol}{A vector of colors with names of species, which are the same as rownames of "phyloData" (may not in the same order).}

\item{showCorVal}{Whether or not show the correlation values, and the default is set as TRUE.}

\item{widthsShinkage}{The shinkage width vector.}

\item{heightsShinkage}{The shinkage width vector.}
}
\value{
A plot object.
}
\description{
Plot correlation matrix of phylogenetic profiles.
}
\details{
A combination plot of correlation matrix with genes and species clusters.
}
\examples{
data(fatp)
ATPCorPlot <- PlotPhyloCor(fatp$atpPhylo, geneCol = fatp$genecol)
\dontrun{
pdf('FATPCorplot.pdf')
ATPCorPlot <- PlotPhyloCor(fatp$atpPhylo, geneCol = fatp$genecol)
dev.off()
}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

