% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PP-class}
\alias{PP-class}
\title{This class represents the raw data structure of phylogenetic profile.}
\description{
This class represents the raw data structure of phylogenetic profile.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{An integer matrix or a numeric matrix, of which the rows are genes/proteins and columns are species. It validates the rownames and colnames of the profile matrix.}
}}

\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
