% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PP.R
\name{PP}
\alias{PP}
\title{Constructor of \code{PP}}
\usage{
PP(x)
}
\arguments{
\item{x}{A numeric matrix.}
}
\value{
A \code{PP} object.
}
\description{
The constructor the \code{PP} class
}
\details{
Construct a \code{PP} object from a numeric matrix.
}
\examples{
require('magrittr')

## construct a PP object without dimnames
ppBinning <- sample(0:1, 10 * 20, replace = TRUE) \%>\% matrix(ncol = 20) \%>\% PP

## with dimnames
ppContinuous <- matrix(rnorm(10 * 20),
                       ncol = 20,
                       dimnames = list(paste0('protein', 1:10),
                                       paste0('spe', 1:20))) \%>\% PP
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
