% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_text.R
\name{pp_text}
\alias{pp_text}
\title{pp plot text}
\usage{
pp_text(x, shift = 0.5, legend.position = "none", ...)
}
\arguments{
\item{x}{A character vector}

\item{shift}{A numeric value indicating shit scale.}

\item{legend.position}{Position of the legend and the default is "none" (no legend). See the \code{theme()} in the ggplot2 package.}

\item{...}{Parameters passed to \code{geom_text()} in the ggplot2 package.}
}
\value{
A \code{gg} class object
}
\description{
Plot text
}
\details{
Plot text, for example column names and row names.
}
\examples{
require(ggplot2)

pp_text(letters[1:10], colour = 'grey55', size = 3)
pp_text(rep(LETTERS[1:2], 3), colour = factor(rep(c('blue', 'red'), 3)))
pp_text(letters[1:5]) + coord_flip()
}
\seealso{
\code{\link[ggplot2]{geom_text}}

\code{\link[ggplot2]{theme}}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
