% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_tile.R
\name{pp_tile}
\alias{pp_tile}
\title{pp plot tiles}
\usage{
pp_tile(x, legend = FALSE, ...)
}
\arguments{
\item{x}{A character vector}

\item{legend}{Whether to contain legends.}

\item{...}{Parameters passed to \code{geom_tile()} in the ggplot2 package.}
}
\value{
A \code{gg} class object
}
\description{
Plot tiles
}
\details{
Plot tiles to discrete groups
}
\examples{
require('ggplot2')

pp_tile(rep(letters[1:3], 4))
pp_tile(rep(LETTERS[1:2], 3)) + scale_fill_manual(values = c('red','blue', 'green'))
pp_tile(rep(letters[1:3], 4)) + coord_flip()
}
\seealso{
\code{\link[ggplot2]{geom_tile}}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
