% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_tree.R
\name{pp_tree}
\alias{pp_tree}
\title{pp plot trees and dendrograms}
\usage{
pp_tree(x, shift = -0.5, ...)
}
\arguments{
\item{x}{A \code{phylo} or a \code{hclust} class object.}

\item{shift}{A numeric value indicating shit scale.}

\item{...}{Parameters passed to \code{geom_segment()} in the ggplot2 package.}
}
\value{
A \code{gg} class object
}
\description{
Plot trees
}
\details{
Plot phylogenetic trees and dendrograms
}
\examples{
require('ggplot2')
require('ape')

treePath <- system.file('extdata', 'bioinfoTree.nex', package = "PhyloProfile")
sceTree <- read.nexus(treePath)

pp_tree(sceTree)
pp_tree(sceTree) + coord_flip() + scale_x_reverse()
pp_tree(sceTree) \%@+\% pp_text(sceTree$tip.label)
}
\seealso{
\code{\link[ggplot2]{geom_segment}}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
