% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/PP.R
\docType{methods}
\name{PPData}
\alias{PPData}
\alias{PPData<-}
\alias{PPData,PP-method}
\alias{PPData<-,PP-method}
\title{Select and replace \code{PP} data}
\usage{
PPData(x, ...)

PPData(x, ...) <- value

\S4method{PPData}{PP}(x, ...)

\S4method{PPData}{PP}(x, ...) <- value
}
\arguments{
\item{x}{A \code{PP} object.}

\item{...}{Additional parameters.}

\item{value}{A numeric matrix.}
}
\value{
A numeric matrix.

\code{PPData(x)}: A numeric matrix.

\code{PPData(x) <- value}: An update PP object.
}
\description{
Select and replace the data matrix of a \code{PP} object
}
\details{
\code{PPData(x)}: Extract the data matrix from a \code{PP} object.

\code{PPData(x) <- value}: Replace the data matrix of a \code{PP} object.
}
\examples{
require('magrittr')
ppContinuous <- matrix(rnorm(10 * 20),
                       ncol = 20,
                       dimnames = list(paste0('protein', 1:10),
                                       paste0('spe', 1:20))) \%>\% PP

## extract data matrix
PPData(ppContinuous)

## replace whole data matrix
rmat <- matrix(sample(0:1, 20 * 30, replace = TRUE),
               ncol = 30,
               dimnames = list(paste0('pro', 1:20),
                               paste0('spe', 1:30)))
PPData(ppContinuous)  <- rmat
## type is changed to "binning"
ppContinuous

}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
