% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRING.R
\name{PPSTRING}
\alias{PPSTRING}
\title{Constructor from the STRING database}
\usage{
PPSTRING(speID, top = 10, symbets = FALSE, STRINGnorm = FALSE, n = 1,
  verbose = FALSE)
}
\arguments{
\item{speID}{Number. NCBI taxonomy identifier must be present in the STRING database. Use \code{get_STRING_species()} from the STRINGdb package to determine supported organisms.}

\item{top}{Number or "all". Top number of proteins or the whole genome (set as "all")}

\item{symbets}{Whether to keep only symmetrical best hits.}

\item{STRINGnorm}{Whether you use the normalized or raw bit scores.}

\item{n}{The number of CPUs or processors, and the default value is 1.}

\item{verbose}{Whether to print retrieve process.}
}
\value{
A \code{PP} object.
}
\description{
The constructor of the \code{PP} class from the STRING database
}
\details{
Retrieve the bit score matrix from a give organism.
}
\examples{
## human top 5 proteins
PPSTRING(9606, top = 5, n = 2)

\dontrun{
## all human proteins
PPRCSTRING(9606, top = "all", n = 4)
}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

