% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PPTreeIdx-class}
\alias{PPTreeIdx-class}
\title{This class represents the data structure of the phylogenetic profile with linkage indices and a phylogenetic tree.}
\description{
This class represents the data structure of the phylogenetic profile with linkage indices and a phylogenetic tree.
}
\section{Slots}{

\describe{
\item{\code{idx}}{An integer matrix with two columns.}
}}

\seealso{
The raw definition of \code{\link[ape]{phylo}}.
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
