% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPTreeIdx.R
\name{PPTreeIdx}
\alias{PPTreeIdx}
\title{Constructor of \code{PPTreeIdx}}
\usage{
PPTreeIdx(pt, x, ...)
}
\arguments{
\item{pt}{A \code{PPTree} object.}

\item{x}{A matrix with two columns or a numeric vector. Proteins not in \code{p} (profile) are removed. The numeric vector indicates the indices of proteins.
\itemize{
  \item a \code{character matrix}: Each row containing the protein names is a linkage.
  \item a \code{numeric matrix} or \code{big.matrix}: Each row containing the protein indices is a linkage.
  \item a \code{numeric vector}: Each element is the index of a interested protein.
}}

\item{...}{Additional parameters if \code{x} is a numeric vector.
\itemize{
  \item \code{y}: Another numeric vector used to generate paired linkages with \code{x}. Each element of \code{x} should be in \code{y}.
  \item \code{self}: Whether include self pairs, and default set is \code{FALSE}.
  \item \code{bidirect}: Whether to include two directions, and default set is \code{FALSE}.
}}
}
\value{
A \code{PPTreeIdx} object.
}
\description{
The constructor the \code{PPTreeIdx} class
}
\details{
Construct a \code{PPTreeIdx} object. The species (columns) are in the same order of phylogenetic tree tips, and for the species not in the tree are deleted.
}
\examples{
require('magrittr')
require('ape')

tree <- rtree(8, tip.label = paste0('spe', 8:1))
ppTree <- sample(0:1, 10 * 20, replace = TRUE) \%>\% matrix(ncol = 20) \%>\% PPTree(tree)

## with self linkages
PPTreeIdx(ppTree, 1:3, 1:3, self = TRUE)
}
\seealso{
PPIdx
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
