% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PP.R, R/PPResult.R
\docType{methods}
\name{[,PP,ANY,ANY,ANY-method}
\alias{[,PP,ANY,ANY,ANY-method}
\alias{[<-,PP,ANY,ANY,numeric-method}
\alias{[,PPResult,ANY,missing,ANY-method}
\title{Select or replace \code{PP}/\code{PPResult} objects}
\usage{
\S4method{[}{PP,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{PP,ANY,ANY,numeric}(x, i, j, ...) <- value

\S4method{[}{PPResult,ANY,missing,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A \code{PP}/\code{PPResult} object.}

\item{i, j, ...}{Indices.}

\item{drop}{Whether the result is coerced to the lowest possible dimension. Desalt set is \code{FALSE}.}

\item{value}{A number, numeric vector or numeric matrix.}
}
\value{
A \code{PP}/\code{PPResult} object.
}
\description{
Select or replace parts of a \code{PP}/\code{PPResult} object
}
\details{
\code{x[i, j, ..., drop]}: Select parts of a \code{PP}/\code{PPResult} object.

\code{x[i, j, ..., drop] <- value}: Replace parts of a \code{PP} object.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
