% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,PP-method}
\alias{show,PP-method}
\title{show methods}
\usage{
\S4method{show}{PP}(object)
}
\arguments{
\item{object}{A \code{PP} instance.}
}
\value{
Show messages.
}
\description{
Show method for \code{PP} instances
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

