% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PP.R, R/PPIdx.R, R/PPResult.R, R/PPTree.R,
%   R/PPTreeIdx.R
\docType{methods}
\name{show,PP-method}
\alias{show,PP-method}
\alias{show,PPIdx-method}
\alias{show,PPResult-method}
\alias{show,PPTree-method}
\alias{show,PPTreeIdx-method}
\title{Show methods}
\usage{
\S4method{show}{PP}(object)

\S4method{show}{PPIdx}(object)

\S4method{show}{PPResult}(object)

\S4method{show}{PPTree}(object)

\S4method{show}{PPTreeIdx}(object)
}
\arguments{
\item{object}{A \code{PP}/\code{PPIdx}/\code{PPTree}/\code{PPTreeIdx}/code{PPResult} object.}
}
\value{
Show messages.
}
\description{
Show method for \code{PP} and \code{PPIdx} objects
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
