% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SimDistBatch.R
\docType{methods}
\name{SimDist}
\alias{SimDist}
\alias{SimDist,PPIdx-method}
\alias{SimDist,PPTreeIdx-method}
\title{Batch process of similarity and distance.}
\usage{
SimDist(x, method, ..., n)

\S4method{SimDist}{PPIdx}(x, method, ..., n = 1)

\S4method{SimDist}{PPTreeIdx}(x, method, ..., n = 1)
}
\arguments{
\item{x}{A \code{PPIdx}/\code{PPTreeIdx} object.}

\item{method}{A character string.
\itemize{
  \item \code{"SimCor"}: Person's correlation coefficient.
  \item \code{"SimJaccard"}: Jaccard similarity.
  \item \code{"SimMI"}: Mutual information.
  \item \code{"DistHamming"}: Hamming distance.
  \item \code{"DistEuclidean"}: Euclidean distance.
}

\code{"SimCor"} and \code{"DistHamming"} are not recommended for binning profiles.}

\item{...}{Additional parameters passed to \code{method}.
\itemize{
  \item \code{bin}: Integer. The number of breaks in the \code{"MI"} method for continuous profiles.
}}

\item{n}{The number of CPUs or processors.}
}
\value{
A numeric vector.
}
\description{
Similarity or distance.
}
\details{
Similarity and distance of paired profiles. If \code{PPTreeIdx} is object is input, the paired profile is collapsed according to the phylogenetic tree.
}
\examples{
require('magrittr')
require('ape')

tree <- system.file('extdata', 'bioinfoTree.nex', package = "PhyloProfile") \%>\% read.nexus
ppPath <- system.file('extdata', 'bioinfoProfile.csv', package = "PhyloProfile")

sceP <- ppPath \%>\% read.csv(row.names = 1) \%>\% as.matrix \%>\% PP
sceL <- PPIdx(sceP, 1:5, 1:5)
sceT <- PPTreeIdx(sceL, tree)

## Mutual information
SimDist(sceL, 'SimMI', n = 2)
SimDist(sceT, 'SimMI', n = 2)

}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
