% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SimDist.R
\docType{methods}
\name{SimDist}
\alias{SimDist}
\alias{SimDist,PPIdx-method}
\title{Batch process of similarity and distance.}
\usage{
SimDist(x, n, method, ...)

\S4method{SimDist}{PPIdx}(x, n = 1, method, ...)
}
\arguments{
\item{x}{A \code{PP} object.}

\item{n}{The number of CPUs or processors.}

\item{method}{A character string. Supported similarity or distance methods:
\itemize{
  \item \code{"SimCor"}: Person's correlation coefficient.
  \item \code{"SimJaccard"}: Jaccard similarity.
  \item \code{"SimMI"}: Mutual information.
  \item \code{"DistHamming"}: Hamming distance.
  \item \code{"DistEuclidean"}: Euclidean distance.
}

\code{"SimCor"} and \code{"DistHamming"} are not recommended for binning profiles.}

\item{...}{Additional parameters passed to \code{method}.
\itemize{
  \item \code{bin}: Integer. The number of breaks in the \code{"MI"} method for continuous profiles.
}}
}
\value{
A numeric vector
}
\description{
Similarity or distance.
}
\details{
SimDist(): Similarity and distance of paired profiles.
}
\examples{
require('magrittr')
data(fatp)

## Mutual information
f1 <- fatp$atpPhylo \%>\% PP \%>\% PPIdx(1:6, 1:6)
SimDist(f1, n = 2, 'SimMI')
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

