% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimMIBin_}
\alias{SimMIBin_}
\alias{SimMIConti_}
\alias{SimCor_}
\alias{SimJaccard_}
\alias{DistHamming_}
\alias{DistManhattan_}
\alias{DistEuclidean_}
\alias{DistMinkowski_}
\title{Similarity and distance}
\usage{
SimMIBin_(f, t)

SimMIConti_(f, t, bin)

SimCor_(f, t)

SimJaccard_(f, t)

DistHamming_(f, t)

DistManhattan_(f, t)

DistEuclidean_(f, t)

DistMinkowski_(f, t, p)
}
\arguments{
\item{f}{Numeric vector indicating a gene profile.}

\item{t}{Numeric vector indicating a gene profile.}

\item{bin}{A positive \code{Integer} indicating the bin.}
}
\value{
A numeric value.
}
\description{
Similarity or distance of paired phylogenetic profile
}
\details{
\code{SimCor_()}: Person's correlation coefficient.

\code{SimJaccard_()}: Jaccard similarity.

\code{SimMIBin_()}: Mutual information for binning data.

\code{SimMIConti_()}: Mutual information for continuous data.

\code{DistHamming_()}: Hamming distance.

\code{DistEuclidean_()}: Euclidean distance.
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
