% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimMIBin}
\alias{DistEuclidean}
\alias{DistHamming}
\alias{SimCor}
\alias{SimJaccard}
\alias{SimMIBin}
\alias{SimMIConti}
\title{similarity and distance}
\usage{
SimMIBin(f, t)

SimMIConti(f, t, bin)

SimCor(f, t)

SimJaccard(f, t)

DistHamming(f, t)

DistEuclidean(f, t)
}
\arguments{
\item{f}{Numeric vector indicating a gene profile.}

\item{t}{Numeric vector indicating a gene profile.}

\item{bin}{Integer.}
}
\value{
A numeric value.
}
\description{
Similarity or distance of paired phylogenetic profile
}
\details{
\code{SimCor()}: Person's correlation coefficient.
\code{SimJaccard()}: Jaccard similarity.
\code{SimMIBin()}: Mutual information for binning data.
\code{SimMIConti()}: Mutual information for continuous data.
\code{DistHamming()}: Hamming distance.
\code{DistEuclidean()}: Euclidean distance.
}
\examples{
## alpha and beta subunits from the F-type ATP synthase.
data(fatp)
a <- t(fatp$atpPhylo['ATP5A1', ])
b <- t(fatp$atpPhylo['ATP5B', ])

## Person's correlation coefficient
corAB <- SimCor(a, b)
## Jaccard similarity
jacAB <- SimJaccard(a, b)
## Mutual information
MIAB <- SimMIBin(a, b)
MIABConti <- SimMIConti(a, b, bin = 10)
## Hamming distance
hamAB <- DistHamming(a, b)
## Eulidean distance
euAB <- DistEuclidean(a, b)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\seealso{
SimDistBatch
}

