% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimCor}
\alias{SimCor}
\alias{SimJaccard}
\alias{DistHamming}
\alias{DistEuclidean}
\title{Similarity and distance}
\usage{
SimCor(f, t)

SimJaccard(f, t)

DistHamming(f, t)

DistEuclidean(f, t)
}
\arguments{
\item{f}{Numeric vector indicating a gene profile.}

\item{t}{Numeric vector indicating a gene profile.}
}
\value{
A numeric value.
}
\description{
Similarity or distance of paired phylogenetic profile
}
\details{
\code{SimCor()}: Person's correlation coefficient.

\code{SimJaccard()}: Jaccard similarity.

\code{SimMIBin()}: Mutual information for binning data.

\code{SimMIConti()}: Mutual information for continuous data.

\code{DistHamming()}: Hamming distance.

\code{DistEuclidean()}: Euclidean distance.
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
