% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimCor_}
\alias{SimCor_}
\alias{SimJaccard_}
\alias{DistHamming_}
\alias{DistManhattan_}
\alias{DistEuclidean_}
\alias{DistMinkowski_}
\title{Similarity and distance}
\usage{
SimCor_(f, t)

SimJaccard_(f, t)

DistHamming_(f, t)

DistManhattan_(f, t)

DistEuclidean_(f, t)

DistMinkowski_(f, t, p)
}
\arguments{
\item{f}{Numeric vector indicating a gene profile.}

\item{t}{Numeric vector indicating a gene profile.}
}
\value{
A numeric value.
}
\description{
Similarity or distance of paired phylogenetic profile
}
\details{
\code{SimCor()}: Person's correlation coefficient.

\code{SimJaccard()}: Jaccard similarity.

\code{SimMIBin()}: Mutual information for binning data.

\code{SimMIConti()}: Mutual information for continuous data.

\code{DistHamming()}: Hamming distance.

\code{DistEuclidean()}: Euclidean distance.
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
