% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDistFun.R
\name{SimCor}
\alias{SimCor}
\alias{SimJaccard}
\alias{SimMI}
\title{Correlation and distance}
\usage{
SimCor(pairProfile)

SimJaccard(pairProfile)

SimMI(pairProfile)
}
\arguments{
\item{pairProfile}{A paired phylogenetic profile. Names of rows are genes and names of columns are species}
}
\value{
A numeric value.
}
\description{
Correlation or distance between a pair of phylogenetic profile
}
\details{
SimCor(): Person's correlation coefficient.
SimJaccard(): Jaccard similarity
SimMI(): Mutual information
}
\examples{
## alpha and beta subunits from the F-type ATP synthase.
data(fatp)
ab <- fatp$atpPhylo[c('ATP5A1', 'ATP5B'), ]

## Person's correlation coefficient
corAB <- SimCor(ab)
## Jaccard similarity
jacAB <- SimJaccard(ab)
## Mutual information
MIAB <- SimMI(ab)

}
\author{
Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}
}

