% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimMIConti}
\alias{DistEuclidean}
\alias{DistHamming}
\alias{SimCor}
\alias{SimJaccard}
\alias{SimMIBin}
\alias{SimMIConti}
\title{similarity and distance}
\usage{
SimMIConti(pairProfile, bin)

SimCor(pairProfile)

SimJaccard(pairProfile)

SimMIBin(pairProfile)

DistHamming(pairProfile)

DistEuclidean(pairProfile)
}
\arguments{
\item{pairProfile}{A paired phylogenetic profile, columns are genes and rows are species.}

\item{bin}{Integer.}
}
\value{
A numeric value.
}
\description{
Similarity or distance of paired phylogenetic profile
}
\details{
SimCor(): Person's correlation coefficient.
SimJaccard(): Jaccard similarity.
SimMIBin(): Mutual information.
SimMIConti(): Mutual information for continuous variables
DistHamming(): Hamming distance.
DistEuclidean(): Euclidean distance.
}
\examples{
## alpha and beta subunits from the F-type ATP synthase.
data(fatp)
ab <- t(fatp$atpPhylo[c('ATP5A1', 'ATP5B'), ])

## Person's correlation coefficient
corAB <- SimCor(ab)
## Jaccard similarity
jacAB <- SimJaccard(ab)
## Mutual information
MIAB <- SimMIBin(ab)
MIABConti <- SimMIConti(ab, bin = 10)
## Hamming distance
hamAB <- DistHamming(ab)
## Eulidean distance
euAB <- DistEuclidean(ab)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\seealso{
SimDistBatch
}

