% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_theme.R
\name{theme_pp}
\alias{theme_pp}
\title{Blank theme}
\usage{
theme_pp(...)
}
\arguments{
\item{...}{Parameters passed to \code{theme()} in the ggplot2 package.}
}
\value{
ggplot2 object
}
\description{
PhyloProfile theme
}
\details{
A totally blank gpplot2 theme expect for the legend
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
