% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utilities.R
\name{Phylo2Mat}
\alias{Phylo2Mat}
\alias{Phylo2MatX}
\alias{Phylo2MatY}
\alias{InitM}
\alias{EachBranch}
\title{tree utilities}
\usage{
Phylo2Mat(x)

Phylo2MatX(x)

Phylo2MatY(x)

InitM(x)

EachBranch(x, m)
}
\arguments{
\item{x}{A \code{phylo} class}

\item{m}{A numeric matrix indicating edge length}

\item{x}{A numeric vector for a single path.}
}
\value{
\code{Phylo2Mat()}: A data.frame indicating x and y locations.

\code{Phylo2MatX()}: A data.frame indicating x locations.

\code{Phylo2MatY()}: A data.frame indicating y locations.

\code{InitM()}: A numeric matrix indicating edge and edge length

\code{EachBranch()}: A numeric matrix indicating y locations for each path.
}
\description{
Deal with \code{phylo} class objects for plot
}
\details{
\code{Phylo2Mat()}: x and y locations.

\code{Phylo2MatX()}: x locations.

\code{Phylo2MatY()}: y locations.

\code{EachBranch()}: y locations for each path.

\code{InitM()}: edge and edge length.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
