% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isBinMat_internal}
\alias{isBinMat_internal}
\title{utilities}
\usage{
isBinMat_internal(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
\code{isBinMat_internal()}: Logic.
}
\description{
Additional utilities
}
\details{
\code{isBinMat_internal()}: Whether a matrix is binning.
}
\keyword{internal}

