% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isBinMat_internal}
\alias{isBinMat_internal}
\alias{validMatNames_internal}
\alias{validTreeMat_internal}
\title{utilities}
\usage{
isBinMat_internal(x)

validMatNames_internal(x, warnName)

validTreeMat_internal(m, t)
}
\arguments{
\item{x}{A matrix.}

\item{warnName}{A character string indicating the class names.}

\item{m}{A numeric matrix.}

\item{t}{A phylo tree.}

\item{x}{A numeric matrix.}
}
\value{
\code{isBinMat_internal()}: Logic.

\code{valiMatNames_internal()}: \code{TRUE} or a warning message.
}
\description{
Additional utilities
}
\details{
\code{isBinMat_internal()}: Whether a matrix is binning.

\code{valiMatNames_internal()}: Validate a numeric matrix with row names and column names.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
